Partially solved final bugs
I have removed redundant functions and variables and made the subjects for the students show in the table. Also, which students the tutor tutors are now editabled and save-able. However, I haven't been abel to save which subjects the tutor teaches each student. I will contiue working on this and on cleaning the code.